Fix overflow/underflow bugs, sign bugs, casting bugs, hopefully all in Temperature / Humidity

+++ software/src/simulation/casts.hpp
@@ -6,6 +6,6 @@ inline constexpr uint8_t u8(auto x) {
   return static_cast<uint8_t>(x);
 }
 
-inline constexpr uint8_t u16(auto x) {
+inline constexpr uint16_t u16(auto x) {
   return static_cast<uint16_t>(x);
 }
